优化代码
将Fragment的代码拆到Presenter
